Fleshed out data buffers: -data buffers are for named values -data buffer arrays are for arrays of values with the same type and format -databufferviews and databufferarrayview are for accessing data in a buffer, but do not own the storage of the buffer to make for cheap copying and partial views
Added working shader constant and instance data uploads using data buffers (not yet being used for actual constant data)
Added shader data format parsing from the scene json
Added shader constants buffer to Models (this will change soon, but they needed a place to be for now)
Moved main loop into a separate class to clean up the main file
Removed instance state value access from ModelInstance because it now can be done through the data buffer directly
Added multiple colour buffers to frame buffers
Added orthographic projection to QuaternionCamera
Made all buffer formats be stored as shared_ptr because nothing owns them definitively so they can't be references anymore
fixed some type warnings

+++ Renderer/Include/Fonts/Font.h
@@ -14,7 +14,7 @@ class Font
 {
 private:
 	float scale;
-	int lineHeight;
+	float lineHeight;
 	std::unique_ptr<stbtt_packedchar_DIDNT_NAME_THIS_STRUCT[]> packedCharData;
 	std::shared_ptr<TextureBuffer> m_bitmap;	
 	std::shared_ptr<Model> m_model;
